Create new Parser Remove unused code, vars & imports
