+++ src/Command/CreateCommand.php
@@ -2,10 +2,8 @@
 
 namespace Pheanstalk\Command;
 
-use Pheanstalk\Exception;
-use Pheanstalk\Response;
+use Pheanstalk\ResponseParser;
 use Pheanstalk\Structure\Workflow;
-use Pheanstalk\XmlResponseParser;
 
 /**
  * The 'Create' command.
@@ -17,7 +15,7 @@ use Pheanstalk\XmlResponseParser;
  * @package Pheanstalk
  * @license http://www.opensource.org/licenses/mit-license.php
  */
-class CreateCommand extends AbstractCommand implements \Pheanstalk\ResponseParser
+class CreateCommand extends AbstractCommand implements ResponseParser
 {
     /** @var Workflow $workflow */
     private $workflow;
