+++ src/Command/CreateTubeCommand.php
@@ -2,11 +2,9 @@
 
 namespace Pheanstalk\Command;
 
-use Pheanstalk\Exception;
-use Pheanstalk\Response;
+use Pheanstalk\ResponseParser;
 use Pheanstalk\Structure\Tube;
 use Pheanstalk\Structure\Workflow;
-use Pheanstalk\XmlResponseParser;
 
 /**
  * The 'CreateTube' command.
@@ -18,7 +16,7 @@ use Pheanstalk\XmlResponseParser;
  * @package Pheanstalk
  * @license http://www.opensource.org/licenses/mit-license.php
  */
-class CreateTubeCommand extends AbstractCommand implements \Pheanstalk\ResponseParser
+class CreateTubeCommand extends AbstractCommand implements ResponseParser
 {
 
     /** @var Tube $tube */
